+++ src/com/maddyhome/idea/vim/command/Command.kt
@@ -20,7 +20,6 @@ package com.maddyhome.idea.vim.command
 
 import com.maddyhome.idea.vim.handler.EditorActionHandlerBase
 import java.util.*
-import javax.swing.KeyStroke
 
 /**
  * This represents a single Vim command to be executed (operator, motion, text object, etc.). It may optionally include
@@ -28,14 +27,17 @@ import javax.swing.KeyStroke
  */
 data class Command(
   var rawCount: Int,
-  var action: EditorActionHandlerBase,
+  var action: EditorActionHandlerBase?,
   val type: Type,
-  var flags: EnumSet<CommandFlags>,
-  var keys: List<KeyStroke>
+  var flags: EnumSet<CommandFlags>
 ) {
 
+  constructor(rawCount: Int, register: Char): this(rawCount, null, Type.SELECT_REGISTER, EnumSet.of(CommandFlags.FLAG_EXPECT_MORE)) {
+    this.register = register
+  }
+
   init {
-    action.process(this)
+    action?.process(this)
   }
 
   var count: Int
@@ -45,6 +47,7 @@ data class Command(
     }
 
   var argument: Argument? = null
+  var register: Char? = null
 
   enum class Type {
     /**
@@ -72,12 +75,12 @@ data class Command(
      */
     COPY,
     PASTE,
-    // TODO REMOVE?
-    RESET,
     /**
      * Represents commands that select the register.
      */
     SELECT_REGISTER,
+    // TODO REMOVE?
+    RESET,
     OTHER_READONLY,
     OTHER_WRITABLE,
     /**
@@ -88,7 +91,7 @@ data class Command(
 
     val isRead: Boolean
       get() = when (this) {
-        MOTION, COPY, SELECT_REGISTER, OTHER_READONLY, COMPLETION -> true
+        MOTION, COPY, OTHER_READONLY, COMPLETION -> true
         else -> false
       }