+++ src/com/maddyhome/idea/vim/command/CommandBuilder.kt
@@ -9,12 +9,13 @@ import java.util.*
 import javax.swing.KeyStroke
 
 class CommandBuilder(private var currentCommandPartNode: CommandPartNode) {
-  private val commandParts = Stack<Command>()
-  private var keys = mutableListOf<KeyStroke>()
+  private val commandParts = ArrayDeque<Command>()
+  private var keyList = mutableListOf<KeyStroke>()
 
   var commandState = CurrentCommandState.NEW_COMMAND
   var count = 0
     private set
+  val keys: Iterable<KeyStroke> get() = keyList
 
   // The argument type for the current command part's action. Kept separate to handle digraphs and characters. We first
   // try to accept a digraph. If we get it, set expected argument type to character and handle the converted key. If we
@@ -24,7 +25,7 @@ class CommandBuilder(private var currentCommandPartNode: CommandPartNode) {
 
   val isReady get() = commandState == CurrentCommandState.READY
   val isBad get() = commandState == CurrentCommandState.BAD_COMMAND
-  val isEmpty get() = commandParts.empty()
+  val isEmpty get() = commandParts.isEmpty()
   val isAtDefaultState get() = isEmpty && count == 0 && expectedArgumentType == null
 
   val isExpectingCount: Boolean
@@ -35,14 +36,23 @@ class CommandBuilder(private var currentCommandPartNode: CommandPartNode) {
     }
 
   fun pushCommandPart(action: EditorActionHandlerBase) {
-    commandParts.push(Command(count, action, action.type, action.flags, keys))
+    commandParts.add(Command(count, action, action.type, action.flags))
     expectedArgumentType = action.argumentType
-    keys = mutableListOf()
+    count = 0
   }
 
-  fun popCommandPart() {
-    commandParts.pop()
-    expectedArgumentType = if (commandParts.size > 0) commandParts.peek().action.argumentType else null
+  fun pushCommandPart(register: Char) {
+    // We will never execute this command, but we need to push something to correctly handle counts on either side of a
+    // select register command part. e.g. 2"a2d2w or even crazier 2"a2"a2"a2"a2"a2d2w
+    commandParts.add(Command(count, register))
+    expectedArgumentType = null
+    count = 0
+  }
+
+  fun popCommandPart(): Command {
+    val command = commandParts.removeLast()
+    expectedArgumentType = if (commandParts.size > 0) commandParts.peekLast().action?.argumentType else null
+    return command
   }
 
   fun fallbackToCharacterArgument() {
@@ -52,16 +62,24 @@ class CommandBuilder(private var currentCommandPartNode: CommandPartNode) {
     expectedArgumentType = Argument.Type.CHARACTER
   }
 
-  fun addKey(keyStroke: KeyStroke) {
-    keys.add(keyStroke)
+  fun addKey(key: KeyStroke) {
+    keyList.add(key)
   }
 
-  fun addCountCharacter(chKey: Char) {
-    count = (count * 10) + (chKey - '0')
+  fun addCountCharacter(key: KeyStroke) {
+    count = (count * 10) + (key.keyChar - '0')
+    // If count overflows and flips negative, reset to 999999999L. In Vim, count is a long, which is *usually* 32 bits,
+    // so will flip at 2147483648. We store count as an Int, which is also 32 bit.
+    // See https://github.com/vim/vim/blob/b376ace1aeaa7614debc725487d75c8f756dd773/src/normal.c#L631
+    if (count < 0) {
+      count = 999999999
+    }
+    addKey(key)
   }
 
   fun deleteCountCharacter() {
     count /= 10
+    keyList.removeAt(keyList.size - 1)
   }
 
   fun setCurrentCommandPartNode(newNode: CommandPartNode) {
@@ -74,7 +92,7 @@ class CommandBuilder(private var currentCommandPartNode: CommandPartNode) {
 
   fun isAwaitingCharOrDigraphArgument(): Boolean {
     if (commandParts.size == 0) return false
-    val argumentType = commandParts.peek().action.argumentType
+    val argumentType = commandParts.peekLast().action?.argumentType
     return argumentType == Argument.Type.CHARACTER || argumentType == Argument.Type.DIGRAPH
   }
 
@@ -88,12 +106,12 @@ class CommandBuilder(private var currentCommandPartNode: CommandPartNode) {
   }
 
   fun completeCommandPart(argument: Argument) {
-    commandParts.peek().argument = argument
+    commandParts.peekLast().argument = argument
     commandState = CurrentCommandState.READY
   }
 
   fun isDuplicateOperatorKeyStroke(key: KeyStroke): Boolean {
-    val action = commandParts.peek()?.action as? DuplicableOperatorAction
+    val action = commandParts.peekLast()?.action as? DuplicableOperatorAction
     return action?.duplicateWith == key.keyChar
   }
 
@@ -102,27 +120,23 @@ class CommandBuilder(private var currentCommandPartNode: CommandPartNode) {
   }
 
   fun buildCommand(): Command {
-    /* Let's go through the command stack and merge it all into one command. At this time there should never
-       be more than two commands on the stack - one is the actual command, and the other would be a motion
-       command argument needed by the first command */
-    var command: Command = commandParts.pop()
-    while (commandParts.size > 0) {
-      val top: Command = commandParts.pop()
-      top.argument = Argument(command)
-      command = top
-    }
-    return fixCommandCounts(command)
-  }
 
-  private fun fixCommandCounts(command: Command): Command {
-    // If we have a command with a motion command argument, both could have their own counts. We need to adjust the
-    // counts, so the motion gets the product of both counts, and the count associated with the command gets reset.
-    // E.g. 3c2w (change 2 words, three times) becomes c6w (change 6 words)
-    if (command.argument?.type === Argument.Type.MOTION) {
-      val motion = command.argument!!.motion
-      motion.count = if (command.rawCount == 0 && motion.rawCount == 0) 0 else command.count * motion.count
+    var command: Command = commandParts.removeFirst()
+    while (commandParts.size > 0) {
+      val next = commandParts.removeFirst()
+      next.count = if (command.rawCount == 0 && next.rawCount == 0) 0 else command.count * next.count
       command.count = 0
+      if (command.type == Command.Type.SELECT_REGISTER) {
+        next.register = command.register
+        command.register = null
+        command = next
+      }
+      else {
+        command.argument = Argument(next)
+        assert(commandParts.size == 0)
+      }
     }
+    expectedArgumentType = null
     return command
   }
 
@@ -130,11 +144,11 @@ class CommandBuilder(private var currentCommandPartNode: CommandPartNode) {
     resetInProgressCommandPart(commandPartNode)
     commandState = CurrentCommandState.NEW_COMMAND
     commandParts.clear()
+    keyList.clear()
     expectedArgumentType = null
   }
 
   fun resetInProgressCommandPart(commandPartNode: CommandPartNode) {
-    keys.clear()
     count = 0
     setCurrentCommandPartNode(commandPartNode)
   }