+++ test/org/jetbrains/plugins/ideavim/action/CommandCountTest.kt
@@ -21,4 +21,23 @@ class CommandCountTest : VimTestCase() {
     typeText(parseKeys("2d3l"))
     myFixture.checkResult("7890")
   }
+
+  // See https://github.com/vim/vim/blob/b376ace1aeaa7614debc725487d75c8f756dd773/src/normal.c#L631
+  fun `test count resets to 999999999L if gets too large`() {
+    configureByText("1")
+    typeText(parseKeys("12345678901234567890<C-A>"))
+    myFixture.checkResult("1000000000")
+  }
+
+  fun `test count select register count operator count motion`() {
+    configureByText("${c}123456789012345678901234567890")
+    typeText(parseKeys("2\"a3d4l")) // Delete 24 characters
+    myFixture.checkResult("567890")
+  }
+
+  fun `test multiple select register counts`() {
+    configureByText("${c}12345678901234567890123456789012345678901234567890")
+    typeText(parseKeys("2\"a2\"b2\"b2d2l")) // Delete 32 characters
+    myFixture.checkResult("345678901234567890")
+  }
 }